refactor(services): centralize monitoring
