+++ services/stt/main.py
@@ -5,7 +5,7 @@ import queue
 import threading
 from typing import AsyncIterator, Iterator
 
-from fastapi import FastAPI, Request, Response, WebSocket, WebSocketDisconnect
+from fastapi import FastAPI, WebSocket, WebSocketDisconnect
 from google.cloud import speech_v1 as speech
 from google.cloud.speech_v1.types import (
     RecognitionConfig,
@@ -14,30 +14,12 @@ from google.cloud.speech_v1.types import (
 )
 from pydantic import BaseModel
 
-from prometheus_client import (
-    CollectorRegistry,
-    CONTENT_TYPE_LATEST,
-    Counter,
-    generate_latest,
-)
+from services.utils import add_monitoring
 
 app = FastAPI(title="FaithEcho STT Service")
 
-# simple Prometheus counter for HTTP requests
-REGISTRY = CollectorRegistry()
-REQUEST_COUNT = Counter(
-    "http_requests_total",
-    "Total HTTP requests",
-    ["method", "endpoint"],
-    registry=REGISTRY,
-)
-
-
-@app.middleware("http")
-async def _count_requests(request: Request, call_next):
-    response = await call_next(request)
-    REQUEST_COUNT.labels(request.method, request.url.path).inc()
-    return response
+# attach Prometheus metrics middleware and endpoint
+add_monitoring(app)
 
 
 class TextChunk(BaseModel):
@@ -117,13 +99,6 @@ async def health() -> dict[str, str]:
 async def ready() -> dict[str, str]:
     """Readiness probe."""
     return {"status": "ready"}
-
-
-@app.get("/metrics")
-async def metrics() -> Response:
-    """Expose Prometheus metrics."""
-    data = generate_latest(REGISTRY)
-    return Response(content=data, media_type=CONTENT_TYPE_LATEST)
 
 
 @app.websocket("/stream")
